Moved login.jsx duplicate into header/login.jsx
