+++ frontend/src/common/Header/Login.jsx
@@ -4,6 +4,22 @@ import firebase from 'firebase';
 
 import './Login.css';
 
+/**
+ * This Method is for getting the authentication token for the current logged in user from the
+ * cookies within your browser. It returns the Authentication token if it exists, otherwise an empty
+ * string
+ *
+ * @returns {string} which is the auth token
+ * @constructor
+ */
+const getAuthToken = () => {
+  const cookieName = 'threaderAuthToken';
+  // Get name followed by anything except a semicolon
+  const cookieString = RegExp(`${cookieName}=[^;]+`).exec(document.cookie);
+  // Return everything after the equal sign, or an empty string if the cookie name not found
+  return decodeURIComponent(cookieString ? cookieString.toString().replace(/^[^=]+./, '') : '');
+};
+
 const Login = ({ handleLogin, displayProfile }) => {
   const uiConfig = {
     signInFlow: 'popup',
@@ -37,4 +53,4 @@ const Login = ({ handleLogin, displayProfile }) => {
   );
 };
 
-export default Login;
+export { Login, getAuthToken };
